DMP-4132 (Dev Only) ARM RPO - Integrate getRecordManagementMatter
Added changes as per code review
