+++ src/main/java/uk/gov/hmcts/darts/arm/client/ArmRpoClient.java
@@ -3,7 +3,7 @@ package uk.gov.hmcts.darts.arm.client;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestHeader;
-import uk.gov.hmcts.darts.arm.model.rpo.RecordManagementMatterResponse;
+import uk.gov.hmcts.darts.arm.client.model.rpo.RecordManagementMatterResponse;
 
 import static org.springframework.http.HttpHeaders.AUTHORIZATION;
 import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
@@ -14,7 +14,7 @@ import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
 )
 public interface ArmRpoClient {
 
-    @PostMapping(value = "${darts.storage.arm-api.rpo-url.get-record-management-matter}",
+    @PostMapping(value = "${darts.storage.arm-api.rpo-url.get-record-management-matter-path}",
         consumes = APPLICATION_JSON_VALUE,
         produces = APPLICATION_JSON_VALUE
     )
