+++ src/main/java/uk/gov/hmcts/darts/arm/client/model/rpo/ExtendedSearchesByMatterResponse.java
@@ -1,7 +1,6 @@
-package uk.gov.hmcts.darts.arm.model.rpo;
+package uk.gov.hmcts.darts.arm.client.model.rpo;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.annotation.JsonRootName;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
@@ -10,6 +9,5 @@ import lombok.NoArgsConstructor;
 @Data
 @NoArgsConstructor
 @EqualsAndHashCode(callSuper = true)
-@JsonRootName(value = "ARMAsyncSearchResponse")
-public class ArmAsyncSearchResponse extends AbstractMatterResponse {
+public class ExtendedSearchesByMatterResponse extends AbstractMatterResponse {
 }
