+++ src/main/java/uk/gov/hmcts/darts/arm/client/model/rpo/MasterIndexFieldByRecordClassSchemaResponse.java
@@ -1,0 +1,13 @@
+package uk.gov.hmcts.darts.arm.client.model.rpo;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+import lombok.NoArgsConstructor;
+
+@JsonIgnoreProperties(ignoreUnknown = true)
+@Data
+@NoArgsConstructor
+@EqualsAndHashCode(callSuper = true)
+public class MasterIndexFieldByRecordClassSchemaResponse extends AbstractMatterResponse {
+}
