+++ src/main/java/uk/gov/hmcts/darts/arm/client/model/rpo/ProfileEntitlementResponse.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.darts.arm.model.rpo;
+package uk.gov.hmcts.darts.arm.client.model.rpo;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.Data;
@@ -9,5 +9,5 @@ import lombok.NoArgsConstructor;
 @Data
 @NoArgsConstructor
 @EqualsAndHashCode(callSuper = true)
-public class ExtendedSearchesByMatterResponse extends AbstractMatterResponse {
+public class ProfileEntitlementResponse extends AbstractMatterResponse {
 }
