+++ src/main/java/uk/gov/hmcts/darts/arm/client/model/rpo/StorageAccountResponse.java
@@ -1,13 +1,11 @@
-package uk.gov.hmcts.darts.arm.model.rpo;
+package uk.gov.hmcts.darts.arm.client.model.rpo;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.Data;
-import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
 @Data
 @NoArgsConstructor
-@EqualsAndHashCode(callSuper = true)
-public class ProfileEntitlementResponse extends AbstractMatterResponse {
+public class StorageAccountResponse {
 }
