+++ src/main/java/uk/gov/hmcts/darts/arm/config/ArmRpoConfigurationProperties.java
@@ -1,0 +1,4 @@
+package uk.gov.hmcts.darts.arm.config;
+
+public class ArmRpoConfigurationProperties {
+}
