--- src/main/java/uk/gov/hmcts/darts/arm/model/rpo/StorageAccountResponse.java
@@ -1,11 +1,0 @@
-package uk.gov.hmcts.darts.arm.model.rpo;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@JsonIgnoreProperties(ignoreUnknown = true)
-@Data
-@NoArgsConstructor
-public class StorageAccountResponse {
-}
