+++ src/main/resources/application.yaml
@@ -285,7 +285,7 @@ darts:
         update-metadata-path: /api/v3/UpdateMetadata
         download-data-path: /api/v1/downloadBlob/{cabinet_id}/{record_id}/{file_id}
       rpo-url:
-        get-record-management-matter: /api/v1/getRecordManagementMatter
+        get-record-management-matter-path: /api/v1/getRecordManagementMatter
 
     dets:
       sas-endpoint: ${DETS_SAS_URL_ENDPOINT:}
